# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Documents Service API
  version: 0.1.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    UserIdHeader:
      type: apiKey
      in: header
      name: X-User-Id
...
security:
  - UserIdHeader: []
paths:
...
  /transactions/{transactionId}/checklists:
    post:
      summary: Assign checklist template to a transaction
      parameters:
        - in: path
          name: transactionId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Created }
...